refactor reporting batch lease recovery route
