+++ src/app/router_registry.py
@@ -229,6 +229,9 @@ from app.routers.proposals import router as proposals_router
 from app.routers.reporting_batch_controls import (
     controls_router as reporting_batch_controls_router,
 )
+from app.routers.reporting_batch_lease_recovery import (
+    recovery_router as reporting_batch_lease_recovery_router,
+)
 from app.routers.reporting_batch_status import status_router as reporting_batch_status_router
 from app.routers.reporting_batch_worker import worker_router as reporting_batch_worker_router
 from app.routers.reporting_batches import batches_router as reporting_batches_router
@@ -432,6 +435,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(reporting_batches_router)
     app.include_router(reporting_batch_status_router)
     app.include_router(reporting_batch_controls_router)
+    app.include_router(reporting_batch_lease_recovery_router)
     app.include_router(reporting_batch_worker_router)
     app.include_router(reporting_schedules_router)
     app.include_router(archive_documents_router)
