+++ src/app/routers/reporting_batch_worker.py
@@ -3,10 +3,8 @@ from typing import Annotated
 from fastapi import APIRouter, Body, Path
 
 from app.contracts.reporting import (
-    BATCH_RECOVERY_RESPONSE_EXAMPLE,
     BATCH_WORKER_RUN_REQUEST_EXAMPLE,
     BATCH_WORKER_RUN_RESPONSE_EXAMPLE,
-    BatchRecoveryResponse,
     BatchWorkerRunRequest,
     BatchWorkerRunResponse,
 )
@@ -16,31 +14,6 @@ from app.routers.reporting_errors import report_batch_error_response
 from app.services.reporting_service_provider import reporting_batch_control_service
 
 worker_router = APIRouter(prefix="/api/v1/report-batches", tags=["Report Batches"])
-
-
-@worker_router.post(
-    "/{batch_id}:recover-expired-leases",
-    response_model=BatchRecoveryResponse,
-    summary="Recover expired report batch item leases",
-    description=(
-        "Recover expired unjobbed item leases through lotus-report so the worker can safely "
-        "redispatch them without duplicating existing report jobs."
-    ),
-    openapi_extra={
-        "responses": {
-            "200": {"content": {"application/json": {"example": BATCH_RECOVERY_RESPONSE_EXAMPLE}}}
-        }
-    },
-)
-async def recover_expired_report_batch_leases(
-    batch_id: Annotated[str, Path(description="Opaque durable report batch identifier.")],
-    caller_headers: ReportingCallerContext,
-) -> BatchRecoveryResponse:
-    return await reporting_batch_control_service().recover_expired_leases(
-        batch_id=batch_id,
-        caller_headers=caller_headers,
-        correlation_id=correlation_id_var.get(),
-    )
 
 
 @worker_router.post(
